Split file listing out of `indexing`
As I'm working to bring back https://github.com/Shopify/saturn/pull/205, it just
makes more sense to have the listing and indexing concepts separate.
