+++ rust/saturn-sys/src/graph_api.rs
@@ -5,7 +5,7 @@ use crate::utils;
 use libc::{c_char, c_void};
 use saturn::model::graph::Graph;
 use saturn::model::ids::DeclarationId;
-use saturn::{indexing, resolution};
+use saturn::{indexing, listing, resolution};
 use std::ffi::CString;
 use std::{mem, ptr};
 
@@ -54,7 +54,7 @@ pub unsafe extern "C" fn sat_index_all(
     let file_paths: Vec<String> = unsafe { utils::convert_double_pointer_to_vec(file_paths, count).unwrap() };
     let mut all_errors = Vec::new();
 
-    let (documents, document_errors) = indexing::collect_file_paths(file_paths);
+    let (documents, document_errors) = listing::collect_file_paths(file_paths);
     all_errors.extend(document_errors);
 
     with_graph(pointer, |graph| {
